tests/fixtures/python_kafka/pyproject.toml
[project]
name = "kafka-consumer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.0.2",
    "elasticsearch>=8.12.0",
    "pydantic>=2.6.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.1.0", "pytest-cov>=5.0.0"]
